Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,21 +1,21 @@
 // Your code here...
 #include <stdio.h>
 int main(){
-    int n1,n2;
+    float n1,n2;
     char c;
-    scanf("%d %d %c",&n1,&n2,&c);
+    scanf("%f %f %c",&n1,&n2,&c);
     switch (c){
         case '+':
-        printf("%d",(n1+n2));
+        printf("%.2f",(n1+n2));
         break;
         case '-':
-        printf("%d",(n1-n2));
+        printf("%.2f",(n1-n2));
         break;
         case '*':
-        printf("%d",(n1*n2));
+        printf("%.2f",(n1*n2));
         break;
         case '/':
-        printf("%d",(n1/n2));
+        printf("%.2f",(n1/n2));
         break;
     }
